requirements.txt
pillow
pytesseract
pylint